mspti用例集问题修复 & mspti cmake warning修复
# message auto-generated for no-merge-commit merge:
merge mspti_comm_deal_graph into master

mspti用例集问题修复 & mspti cmake warning修复

Created-by: msprofDeveloper

Match-id-ea46fa8a7695f2ffb1fa568c51e8b7e5b49d94a7

+++ mspti/common/context_manager.cpp
@@ -172,7 +172,7 @@ std::vector<uint64_t> ContextManager::GetRealTimeFromSysCnt(uint32_t deviceId, c
         devTimeInfo = *iter->second;
     }
     std::vector<uint64_t> ans(sysCnts.size());
-    for (int i = 0; i < sysCnts.size(); i++) {
+    for (size_t i = 0; i < sysCnts.size(); i++) {
         ans[i] = CalculateRealTime(sysCnts[i], devTimeInfo);
     }
     return ans;
